ProposedMigration: Explain excuses and suggest course of actions
Expands the proposed migration documentation to include discussion of a
variety of issues that can come up, and techniques for resolving them.

Notably, this further fills in the glossary of terminologies shown on
the Update Excuses page, listed in a way that should be searchable
in-doc as well as linkable via html anchors.  For example, to understand
what "Implicit dependency ..." means, one could do a text search on the
file from the web browser, or direct link to the relevant section via
url like:

   http://umh.com/ProposedMigration.html#implicit-dependency-aaa-bbb

The top level README is also updated with direct links to this and other
pages.