version-strings: Make wording more succinct and explicit
Current wording states that Stable Release Updates have a different
scheme in the "ubuntu revision" section of the version scheme, but does
not explicitly state that the details in this section completely override
other guidance in other sections of the handbook regarding "ubuntu revisions".

Previous wording leaves unclear whether an Ubuntu-specific delta released in a
Stable Release Update should increment a) the Ubuntu delta number (ubuntuN),
which remains part of the "ubuntu revisions" section, or b) the "increment" in
the Stable Release Updates version string.

Also make other wording in this section more direct.

Fixes #97